requirements.txt
appdirs==1.4.4
asgiref==3.4.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django
django-cors-headers==3.4.0
#django-pyodbc-azure==2.1.0.0
django-redis==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.2
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pyjokes==0.6.0
#pyodbc==4.0.31
pytz==2020.5
redis==3.5.3
regex==2021.4.4
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
